Updated PEPE partial charge code so that the aromaticity flags on the atoms and bonds are maintained on return from the function (since the code appears to wipe out these flags). This fixes bug 2013754. Also modified the code to make the gasteiger iterations variable an integer rather than double
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/cdk/trunk@11521 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/test/org/openscience/cdk/charges/GasteigerPEPEPartialChargesTest.java
@@ -212,7 +212,7 @@ public class GasteigerPEPEPartialChargesTest extends NewCDKTestCase {
     public void testSetMaxGasteigerIters_Double() throws Exception {
 
         GasteigerPEPEPartialCharges peoe = new GasteigerPEPEPartialCharges();
-        double MX_ITERATIONS = 10;
+        int MX_ITERATIONS = 10;
         peoe.setMaxGasteigerIters(MX_ITERATIONS);
         Assert.assertEquals(MX_ITERATIONS, peoe.getMaxGasteigerIters());
